+++ src/components/dropdown/Dropdown.vue
@@ -205,7 +205,12 @@ export default {
             if (!this.isActive) {
                 // if not active, toggle after clickOutside event
                 // this fixes toggling programmatic
-                this.$nextTick(() => { this.isActive = !this.isActive })
+                this.$nextTick(() => {
+                    const value = !this.isActive
+                    this.isActive = value
+                    // Vue 2.6.x ???
+                    setTimeout(() => (this.isActive = value))
+                })
             } else {
                 this.isActive = !this.isActive
             }