+++ src/components/toast/index.js
@@ -41,5 +41,5 @@ use(Plugin)
 export default Plugin
 
 export {
-    ToastProgrammatic as Toast
+    ToastProgrammatic
 }